% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Log_optA.R
\name{Log_optA}
\alias{Log_optA}
\title{Title}
\usage{
Log_optA(X, y, beta.prop, idx.prop, PI.prop)
}
\arguments{
\item{X}{Covariate matrix}

\item{y}{Response vector}

\item{PI.prop}{}
}
\value{

}
\description{
Title
}
