% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{uniform.Samp}
\alias{uniform.Samp}
\title{Title}
\usage{
uniform.Samp(X, y, r, intercept = F)
}
\arguments{
\item{X}{Covariate}

\item{y}{Response}

\item{r}{Sample size}

\item{intercept}{Wheter contains intercept for regression}
}
\value{

}
\description{
Title
}
